+++ crates/swc_ecma_transforms_base/src/rename/analyzer/scope.rs
@@ -88,7 +88,7 @@ impl Scope {
         });
     }
 
-    pub(crate) fn rename_single_thread<R>(
+    pub(crate) fn rename_in_normal_mode<R>(
         &mut self,
         renamer: &R,
         to: &mut RenameMap,
@@ -102,7 +102,7 @@ impl Scope {
 
         // let mut cloned_reverse = reverse.clone();
 
-        self.rename_one_scope_single_thread(
+        self.rename_one_scope_in_normal_mode(
             renamer,
             to,
             previous,
@@ -112,7 +112,7 @@ impl Scope {
         );
 
         for child in &mut self.children {
-            child.rename_single_thread(
+            child.rename_in_normal_mode(
                 renamer,
                 to,
                 &Default::default(),
@@ -122,7 +122,7 @@ impl Scope {
         }
     }
 
-    fn rename_one_scope_single_thread<R>(
+    fn rename_one_scope_in_normal_mode<R>(
         &self,
         renamer: &R,
         to: &mut RenameMap,
@@ -187,7 +187,7 @@ impl Scope {
     }
 
     #[cfg_attr(not(feature = "concurrent-renamer"), allow(unused))]
-    pub(crate) fn rename_parallel<R>(
+    pub(crate) fn rename_in_mangle_mode<R>(
         &mut self,
         renamer: &R,
         to: &mut RenameMap,
@@ -203,7 +203,7 @@ impl Scope {
 
         let mut cloned_reverse = reverse.clone();
 
-        self.rename_one_scope_parallel(
+        self.rename_one_scope_in_mangle_mode(
             renamer,
             to,
             previous,
@@ -225,7 +225,7 @@ impl Scope {
                     use std::collections::HashMap;
 
                     let mut new_map = HashMap::default();
-                    child.rename_parallel(
+                    child.rename_in_mangle_mode(
                         renamer,
                         &mut new_map,
                         to,
@@ -245,7 +245,7 @@ impl Scope {
         }
 
         for child in &mut self.children {
-            child.rename_parallel(
+            child.rename_in_mangle_mode(
                 renamer,
                 to,
                 &Default::default(),
@@ -257,7 +257,7 @@ impl Scope {
         }
     }
 
-    fn rename_one_scope_parallel<R>(
+    fn rename_one_scope_in_mangle_mode<R>(
         &self,
         renamer: &R,
         to: &mut RenameMap,